+++ website_event_attendee_fields/tests/common.py
@@ -1,0 +1,27 @@
+from datetime import datetime, timedelta
+
+from odoo import fields
+from odoo.tests.common import TransactionCase
+
+
+class TestCase(TransactionCase):
+    at_install = True
+    post_install = True
+
+    def setUp(self):
+        super(TestCase, self).setUp()
+        self.event = self.env['event.event'].create({
+            'name': 'TestEvent',
+            'attendee_signup': True,
+            'create_partner': True,
+            'date_begin': fields.Datetime.to_string(datetime.today() + timedelta(days=1)),
+            'date_end': fields.Datetime.to_string(datetime.today() + timedelta(days=15)),
+        })
+        self.event.write({
+            'attendee_field_ids': [(6, 0, [
+                self.env.ref('website_event_attendee_fields.attendee_field_name').id,
+                self.env.ref('website_event_attendee_fields.attendee_field_email').id,
+                self.env.ref('website_event_attendee_fields.attendee_field_phone').id,
+                self.env.ref('website_event_attendee_fields.attendee_field_country_id').id,
+            ])]
+        })