+++ website_multi_company/controllers/main.py
@@ -1,4 +1,6 @@
-# -*- coding: utf-8 -*-
+# Copyright 2018 Ivan Yelizariev <https://it-projects.info/team/yelizariev>
+# Copyright 2018 Ildar Nasyrov <https://it-projects.info/team/iledarn>
+# License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 from odoo.http import request
 from odoo.addons.website_sale.controllers.main import WebsiteSale
 
@@ -10,4 +12,4 @@ class WebsiteSaleExtended(WebsiteSale):
         company = request.website.company_id
         if not company:
             return domain
-        return [('company_id', '=', company.id)] + domain
+        return ['|', ('company_id', '=', company.id), ('company_id', '=', False)] + domain