+++ website_multi_company_portal/controllers/main.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from odoo.http import request
 from odoo.addons.portal.controllers.portal import CustomerPortal