+++ website_sale_search_clear/controllers.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from odoo import http
 from odoo.addons.website_sale.controllers.main import WebsiteSale as controller
 
@@ -9,5 +8,4 @@ class WebsiteSale(controller):
     def shop(self, page=0, category=None, search='', ppg=False, **post):
         if category and search:
             category = None
-        return super(WebsiteSale, self).shop(page, category, search, ppg,
-                                             **post)
+        return super(WebsiteSale, self).shop(page, category, search, ppg, **post)