+++ website_sale_stock_status/__init__.py
@@ -1,2 +1,1 @@
-# -*- coding: utf-8 -*-
 from . import website_sale_stock_status_models