+++ website_multi_company_sale/doc/changelog.rst
@@ -1,3 +1,13 @@
+`1.2.0`
+-------
+
+- **ADD:** you can now specify allowed websites for your payment acquirers
+
+`1.1.0`
+-------
+
+- **ADD:** you can now specify allowed websites for your products
+
 `1.0.2`
 -------